refactor: move DialectUnparser to query.rs and box large enum variant
Move DialectUnparser and SqlDialect::unparser() from spark_dialect.rs
to query.rs to co-locate with the SqlDialect enum. Box the Spark
CustomDialect variant to fix clippy::large_enum_variant lint.
